+++ Changelog.md
@@ -16,6 +16,7 @@
 xxx
 ### Changed
 - when rescan the tag a new random playlist is generated when the player is closed
+- added option in folder to play new file on rescan
 
 ## [0.17.0]
 ### Added